@@ -14,7 +14,7 @@
   ],
   "require": {
     "pear/archive_tar": "^1.4.14",
-    "php": ">=7.4",
+    "php": ">=8.0",
     "ext-curl": "*",
     "ext-zip": "*",
     "ext-mbstring": "*",
@@ -22,8 +22,7 @@
     "symfony/console": "^5.4"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.0",
-    "sebastian/phpcpd": "^6.0",
+    "phpunit/phpunit": "^9.6",
     "squizlabs/php_codesniffer": "^3.5",
     "donatj/mock-webserver": "^2.6",
     "phpstan/phpstan": "^1.4"
@@ -40,21 +39,21 @@
   },
   "scripts": {
     "phpstan": [
-      "php -d memory_limit=-1 ./vendor/bin/phpstan analyze -l 3 src tests"
+      "php -d memory_limit=-1 ./vendor/bin/phpstan analyze -l 7 src tests"
     ],
     "profile": [
       "php -d xdebug.mode=profile -d xdebug.output_dir=mytracedir/ -d xdebug.start_with_request=yes -d xdebug.use_compression=true ./vendor/bin/phpunit"
     ],
     "check": [
-      "./vendor/bin/phpcs --standard=PSR12 src tests",
-      "./vendor/bin/phpcpd --suffix='.php' src"
+      "./vendor/bin/phpcs --standard=PSR12 src tests"
     ],
     "phpunit": [
-      "phpdbg -qrr ./vendor/bin/phpunit -d memory_limit=-1 --testsuite BagIt"
+      "phpdbg -qrr ./vendor/bin/phpunit -d memory_limit=-1 --verbose --testsuite BagIt"
     ],
     "test": [
       "@check",
-      "@phpunit"
+      "@phpunit",
+      "@phpstan"
     ]
   },
   "config": {
