@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace whikloj\BagItTools\Test;
 
+use Exception;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
+use ReflectionException;
 use ReflectionMethod;
 use whikloj\BagItTools\Bag;
 use whikloj\BagItTools\BagUtils;
@@ -70,7 +72,7 @@ class BagItTestFramework extends TestCase
      * the test throws an exception to ensure it gets deleted.
      * @var string
      */
-    protected $tmpdir;
+    protected string $tmpdir;
 
     /**
      * {@inheritdoc}
@@ -87,7 +89,7 @@ class BagItTestFramework extends TestCase
     public function tearDown(): void
     {
         parent::tearDown();
-        if (isset($this->tmpdir) && file_exists($this->tmpdir)) {
+        if (file_exists($this->tmpdir)) {
             self::deleteDirAndContents($this->tmpdir);
         }
     }
@@ -97,7 +99,7 @@ class BagItTestFramework extends TestCase
      *
      * @return string
      *   The filename.
-     * @throws \whikloj\BagItTools\Exceptions\FilesystemException
+     * @throws FilesystemException
      *   Unable to generate a new temporary directory.
      */
     protected function getTempName(): string
@@ -116,11 +118,17 @@ class BagItTestFramework extends TestCase
      *
      * @param string $path
      *   The directory to delete.
+     * @throws Exception
+     *   Unable to scan directory contents.
      */
     protected static function deleteDirAndContents(string $path): void
     {
         if (is_dir($path)) {
-            $files = array_diff(scandir($path), [".", ".."]);
+            $dir_files = scandir($path);
+            if ($dir_files === false) {
+                throw new Exception("Unable to read directory contents");
+            }
+            $files = array_diff($dir_files, [".", ".."]);
             foreach ($files as $file) {
                 $currentFile = $path . '/' . $file;
                 if (is_dir($currentFile)) {
@@ -161,6 +169,8 @@ class BagItTestFramework extends TestCase
      *   The source directory.
      * @return string
      *   The path to the copy of the bag.
+     * @throws FilesystemException
+     *   Unable to create temporary directory.
      */
     protected function copyTestBag(string $testDir): string
     {
@@ -173,8 +183,8 @@ class BagItTestFramework extends TestCase
     /**
      * Compare two arrays have all the same elements, does not compare order.
      *
-     * @param array $expected The expected array.
-     * @param array $testing The array to test.
+     * @param array<int|string, int|string|array<string>> $expected The expected array.
+     * @param array<int|string, int|string|array<string>> $testing The array to test.
      */
     protected function assertArrayEquals(array $expected, array $testing): void
     {
@@ -191,10 +201,15 @@ class BagItTestFramework extends TestCase
      *
      * @param string $src The original directory.
      * @param string $dest The destination directory.
+     * @throws Exception Unable to scan directory contents.
      */
     private static function copyDir(string $src, string $dest): void
     {
-        $files = array_diff(scandir($src), [".", ".."]);
+        $dir_files = scandir($src);
+        if ($dir_files === false) {
+            throw new Exception("Unable to read directory contents");
+        }
+        $files = array_diff($dir_files, [".", ".."]);
         foreach ($files as $item) {
             if (is_dir("$src/$item")) {
                 if (!is_dir("$dest/$item")) {
@@ -210,15 +225,15 @@ class BagItTestFramework extends TestCase
     /**
      * Get a private or protected method to test it directly.
      *
-     * @param string $class
+     * @param class-string $class
      *   Class to refect.
      * @param string $method
      *   Method to get.
      *
-     * @return \ReflectionMethod
+     * @return ReflectionMethod
      *   Reflection of the method.
      *
-     * @throws \ReflectionException
+     * @throws ReflectionException
      */
     protected static function getReflectionMethod(string $class, string $method): ReflectionMethod
     {
@@ -230,7 +245,7 @@ class BagItTestFramework extends TestCase
 
     /**
      * Assert the encoding in the bagit.txt is X and the version is 1.0
-     * @param \whikloj\BagItTools\Bag $bag
+     * @param Bag $bag
      *   The bag.
      * @param string $version_string
      *   The BagIt version.
@@ -242,7 +257,7 @@ class BagItTestFramework extends TestCase
 
     /**
      * Assert the encoding in the bagit.txt is X and the version is 1.0
-     * @param \whikloj\BagItTools\Bag $bag
+     * @param Bag $bag
      *   The bag.
      * @param string $encoding
      *   The file encoding.
@@ -254,7 +269,7 @@ class BagItTestFramework extends TestCase
 
     /**
      * Assert the version and encoding in the actual bagit.txt on disk
-     * @param \whikloj\BagItTools\Bag $bag
+     * @param Bag $bag
      *   The bag.
      * @param string|null $version_string
      *   The version string or null to use the default.
@@ -292,11 +307,46 @@ class BagItTestFramework extends TestCase
     protected function assertStringContainsStringWithoutNewlines(string $expected, string $original): void
     {
         $split_original = preg_split("/(\r\n|\r|\n)/", $original);
+        if ($split_original === false) {
+            $this->fail("Unable to split string into lines");
+        }
         array_walk($split_original, function (&$o) {
             $o = trim($o);
         });
         $new_original = implode(" ", $split_original);
         $final = trim($new_original);
         $this->assertStringContainsString($expected, $final);
+    }
+
+    /**
+     * Get the current working directory.
+     *
+     * @return string
+     *   The current working directory.
+     * @throws Exception
+     *   Unable to get the current working directory.
+     */
+    protected static function getCwd(): string
+    {
+        $cwd = getcwd();
+        if ($cwd === false) {
+            throw new Exception("Unable to get current working directory");
+        }
+        return $cwd;
+    }
+
+    /**
+     * @param string $path The path to the file.
+     * @param string $mode The mode to open the file in.
+     * @return resource The file pointer.
+     * @throws Exception Unable to open the file.
+     */
+    protected static function openFile(string $path, string $mode)
+    {
+        $fp = fopen($path, $mode);
+        if ($fp === false) {
+            throw new Exception("Unable to open file " . basename($path));
+        }
+        return $fp;
     }
 }
