@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace whikloj\BagItTools;
 
+use whikloj\BagItTools\Exceptions\FilesystemException;
+
 /**
  * Payload Manifest extension of AbstractManifest.
  *
@@ -16,13 +18,13 @@ class PayloadManifest extends AbstractManifest
     /**
      * PayloadManifest constructor.
      *
-     * @param \whikloj\BagItTools\Bag $bag
+     * @param Bag $bag
      *   The bag this manifest is part of.
      * @param string $algorithm
      *   The BagIt name of the hash algorithm.
      * @param bool $load
      *   Whether we are loading an existing file
-     * @throws \whikloj\BagItTools\Exceptions\FilesystemException
+     * @throws FilesystemException
      *   Unable to read manifest file.
      */
     public function __construct(Bag $bag, string $algorithm, bool $load = false)
