@@ -51,6 +51,9 @@ class ExtendedBagTest extends BagItTestFramework
         $this->assertFalse($bag->isExtended());
         $payloads = array_keys($bag->getPayloadManifests());
         $hash = reset($payloads);
+        if ($hash === false) {
+            $this->fail("No payload manifest found.");
+        }
         $manifests = $bag->getTagManifests();
         $this->assertCount(0, $manifests);
 
@@ -418,7 +421,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testSetBagInfoElement(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
         $this->assertCount(1, $bag->getBagInfoData());
         $this->assertTrue($bag->hasBagInfoTag('contact-name'));
@@ -452,7 +454,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testSetGeneratedField(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $bag->addBagInfoTag('Source-organization', 'Planet Earth');
         // Doesn't match due to underscore instead of hyphen.
         $bag->addBagInfoTag('PAYLOAD_OXUM', '123456.12');
@@ -474,7 +475,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testAddBagInfoTags(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $this->assertArrayEquals([], $bag->getBagInfoData());
         $baginfo = $bag->getBagRoot() . '/bag-info.txt';
 
@@ -504,7 +504,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testAddBagInfoTagsMultiple(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $this->assertArrayEquals([], $bag->getBagInfoData());
         $baginfo = $bag->getBagRoot() . '/bag-info.txt';
 
@@ -536,7 +535,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testAddBagInfoTagsGenerated(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $this->assertArrayEquals([], $bag->getBagInfoData());
 
         $inputTags = [
@@ -551,13 +549,14 @@ class ExtendedBagTest extends BagItTestFramework
     }
 
     /**
-     * Test that addBagInfoTags() throws an exception if the bag is not extended.
+     * Test that addBagInfoTags() no longer throws an exception if the bag is not extended.
      * @group Extended
      * @covers ::addBagInfoTags
      */
     public function testAddBagInfoTagsNotExtended(): void
     {
         $bag = Bag::create($this->tmpdir);
+        $this->assertFalse($bag->isExtended());
         $inputTags = [
             'Source-organization' => 'The Pyramid',
             'CONTACT-NAME' => [
@@ -565,9 +564,12 @@ class ExtendedBagTest extends BagItTestFramework
                 'Monty Hall',
             ],
         ];
-        $this->expectException(BagItException::class);
-        $this->expectExceptionMessage("This bag is not extended, you need '\$bag->setExtended(true);'");
         $bag->addBagInfoTags($inputTags);
+        $this->assertTrue($bag->isExtended());
+        $this->assertTrue($bag->hasBagInfoTag('source-organization'));
+        $this->assertTrue($bag->hasBagInfoTag('contact-name'));
+        $this->assertArrayEquals(['The Pyramid'], $bag->getBagInfoByTag('source-organization'));
+        $this->assertArrayEquals(['Monty Hall', 'Bob Barker'], $bag->getBagInfoByTag('contact-name'));
     }
 
     /**
@@ -622,7 +624,6 @@ class ExtendedBagTest extends BagItTestFramework
         $this->tmpdir = $this->prepareBasicTestBag();
         $bag = Bag::load($this->tmpdir);
         $payloads = $bag->getPayloadManifests();
-        $this->assertTrue(is_array($payloads));
         $this->assertCount(1, $payloads);
         $tags = $bag->getTagManifests();
         $this->assertCount(0, $tags);
@@ -640,7 +641,6 @@ class ExtendedBagTest extends BagItTestFramework
         $this->tmpdir = $this->prepareExtendedTestBag();
         $bag = Bag::load($this->tmpdir);
         $payloads = $bag->getPayloadManifests();
-        $this->assertTrue(is_array($payloads));
         $this->assertCount(1, $payloads);
         $tags = $bag->getTagManifests();
         $this->assertTrue(is_array($tags));
@@ -711,7 +711,6 @@ class ExtendedBagTest extends BagItTestFramework
     public function testLongBagInfoLinesWrap(): void
     {
         $bag = Bag::create($this->tmpdir);
-        $bag->setExtended(true);
         $title = 'A really long long long long long long long long long long long title'
             . ' with a colon : between and more information are on the way';
         $bag->addBagInfoTag('Title', $title);
@@ -827,9 +826,13 @@ class ExtendedBagTest extends BagItTestFramework
         ];
         foreach ($files as $file) {
             $path = $this->tmpdir . '/' . $file;
+            $contents = file_get_contents($path);
+            if ($contents === false) {
+                $this->fail("Failed to read file: $path");
+            }
             file_put_contents(
                 $path,
-                str_replace("\n", $newEnding, file_get_contents($path))
+                str_replace("\n", $newEnding, $contents)
             );
         }
     }
@@ -991,5 +994,298 @@ class ExtendedBagTest extends BagItTestFramework
         );
         $bag = Bag::load($this->tmpdir);
         $this->assertTrue($bag->isValid());
+    }
+
+    /**
+     * Test trying to add a tag file with a non-existant file.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagFileFileDoesNotExist(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("/path/to/nowhere.txt does not exist, is not a file or is not readable.");
+        $bag->addTagFile('/path/to/nowhere.txt', 'nowhere.txt');
+    }
+
+    /**
+     * Test trying to add a tag file over the bag-info.txt file.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagFileCannotOverwriteBagInfo(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter reserved file (bag-info.txt) file with your own tag file.");
+        $bag->addTagFile($special_file, 'bag-info.txt');
+    }
+
+    /**
+     * Test trying to add a tag file over the bagit.txt file.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagCannotOverwriteBagIt(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter reserved file (bagit.txt) file with your own tag file.");
+        $bag->addTagFile($special_file, 'bagit.txt');
+    }
+
+    /**
+     * Test trying to add a tag file over a payload manifest file.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagCannotOverwritePayloadManifest(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $bag->addAlgorithm('sha256');
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter a manifest or tag manifest file with your own tag file.");
+        $bag->addTagFile($special_file, 'manifest-sha256.txt');
+    }
+
+    /**
+     * Test trying to add a tag file over a tag manifest file.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagCannotOverwriteTagManifest(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->setExtended(true);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $bag->addAlgorithm('sha256');
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter a manifest or tag manifest file with your own tag file.");
+        $bag->addTagFile($special_file, 'tagmanifest-sha256.txt');
+    }
+
+    /**
+     * Test trying to add a tag file to the data directory.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagCannotBeInDataDirectory(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag files must be in the bag root or a tag file directory");
+        $bag->addTagFile($special_file, 'data/special.txt');
+    }
+
+    /**
+     * Test trying to add a tag file to the data directory.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagCannotBeInDataDirectory2(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag files must be in the bag root or a tag file directory");
+        $bag->addTagFile($special_file, './data/special.txt');
+    }
+
+    /**
+     * Test adding a tag file to the bag.
+     * @group Extended
+     * @covers ::addTagFile
+     */
+    public function testAddTagNotOverwrite(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addTagFile($special_file, 'special.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special.txt');
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag file (special.txt) already exists in the bag.");
+        $bag->addTagFile($special_file, 'special.txt');
+    }
+
+    /**
+     * Test adding a tag file to the bag.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::checkTagFileConstraints
+     */
+    public function testAddTagFileSuccess(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addTagFile($special_file, 'special.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special.txt');
+    }
+
+    /**
+     * Test adding tag files in a subdirectory
+     * @group Extended
+     * @covers ::addTagFile
+     */
+    public function testAddTagFileInSubDir(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addTagFile($special_file, 'special/special.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special/special.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special');
+        $bag->addTagFile($special_file, 'special/special2.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special/special2.txt');
+    }
+
+    /**
+     * Test trying to add a tag file with a non-existant file.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagFileFileDoesNotExist(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag file (nowhere.txt) does not exist in the bag.");
+        $bag->removeTagFile('nowhere.txt');
+    }
+
+    /**
+     * Test trying to remove the bag-info.txt file using removeTagFile.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagFileCannotOverwriteBagInfo(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $bag->update(); // Update to create the bag-info.txt on disk.
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/bag-info.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter reserved file (bag-info.txt) file with your own tag file.");
+        $bag->removeTagFile('bag-info.txt');
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/bag-info.txt'));
+    }
+
+    /**
+     * Test trying to remove the bagit.txt file using removeTagFile.
+     * @group Extended
+     * @covers ::addTagFile
+     */
+    public function testRemoveTagCannotOverwriteBagIt(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/bagit.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter reserved file (bagit.txt) file with your own tag file.");
+        $bag->removeTagFile('bagit.txt');
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/bagit.txt'));
+    }
+
+    /**
+     * Test trying to remove a payload manifest file using removeTagFile.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagCannotOverwritePayloadManifest(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $bag->setAlgorithm('sha256');
+        $bag->update();
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/manifest-sha256.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter a manifest or tag manifest file with your own tag file.");
+        $bag->removeTagFile('manifest-sha256.txt');
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/manifest-sha256.txt'));
+    }
+
+    /**
+     * Test trying to remove a tag manifest file using removeTagFile.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagCannotOverwriteTagManifest(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $bag->addBagInfoTag('Contact-NAME', 'Monty Hall');
+        $bag->setAlgorithm('sha256');
+        $bag->update();
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/tagmanifest-sha256.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("You cannot alter a manifest or tag manifest file with your own tag file.");
+        $bag->removeTagFile('tagmanifest-sha256.txt');
+        $this->assertTrue(file_exists($bag->getBagRoot() . '/tagmanifest-sha256.txt'));
+    }
+
+    /**
+     * Test trying to remove a file from the data directory using removeTagFile.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagCannotBeInDataDirectory(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $bag->addFile(self::TEST_RESOURCES . '/text/empty.txt', 'data/empty.txt');
+        $this->assertTrue(file_exists($bag->getDataDirectory() . '/empty.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag files must be in the bag root or a tag file directory");
+        $bag->removeTagFile('data/special.txt');
+        $this->assertTrue(file_exists($bag->getDataDirectory() . '/empty.txt'));
+    }
+
+    /**
+     * Test trying to remove a file from the data directory using removeTagFile.
+     * @group Extended
+     * @covers ::removeTagFile
+     */
+    public function testRemoveTagCannotBeInDataDirectory2(): void
+    {
+        $bag = Bag::create($this->tmpdir);
+        $bag->addFile(self::TEST_RESOURCES . '/text/empty.txt', 'data/empty.txt');
+        $this->assertTrue(file_exists($bag->getDataDirectory() . '/empty.txt'));
+        $this->expectException(BagItException::class);
+        $this->expectExceptionMessage("Tag files must be in the bag root or a tag file directory");
+        $bag->removeTagFile('./data/special.txt');
+        $this->assertTrue(file_exists($bag->getDataDirectory() . '/empty.txt'));
+    }
+
+    /**
+     * Test adding and removing multiple tag files in a tag file subdirectory.
+     * @group Extended
+     * @covers ::addTagFile
+     * @covers ::removeTagFile
+     */
+    public function testAddAndRemoveTagFileInSubDir(): void
+    {
+        $special_file = self::TEST_RESOURCES . '/tagFiles/SomeSpecialTags.txt';
+        $bag = Bag::create($this->tmpdir);
+        $bag->addTagFile($special_file, 'special/special.txt');
+        $bag->addTagFile($special_file, 'special/special2.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special/special.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special/special2.txt');
+        $bag->removeTagFile('special/special2.txt');
+        $this->assertFileExists($bag->getBagRoot() . '/special/special.txt');
+        $this->assertFileDoesNotExist($bag->getBagRoot() . '/special/special2.txt');
+        $bag->removeTagFile('special/special.txt');
+        $this->assertFileDoesNotExist($bag->getBagRoot() . '/special/special.txt');
+        $this->assertFileDoesNotExist($bag->getBagRoot() . '/special/special2.txt');
+        $this->assertFileDoesNotExist($bag->getBagRoot() . '/special');
     }
 }
